+++ util/SerializeUniverse.cpp
@@ -260,8 +260,8 @@ void SpeciesManager::serialize(Archive& ar, const unsigned int version)
     // empires and eachother
 
     std::map<std::string, std::set<int> >                   species_homeworlds;
-    std::map<std::string, std::map<int, double> >           empire_opinions;
-    std::map<std::string, std::map<std::string, double> >   other_species_opinions;
+    std::map<std::string, std::map<int, float> >            empire_opinions;
+    std::map<std::string, std::map<std::string, float> >    other_species_opinions;
     std::map<std::string, std::map<int, float> >            species_object_populations;
     std::map<std::string, std::map<std::string, int> >      species_ships_destroyed;
